Move info button to stable directory and build
