+++ src/stable/components/InfoButton/InfoButton.js
@@ -1,6 +1,6 @@
 import styles from '!!raw-loader!./InfoButton.css';
-import varStyles from '!!raw-loader!../../../../shared/variables.css';
-import bootstrapStyles from '!!raw-loader!../../../../shared/themed-bootstrap.css';
+import varStyles from '!!raw-loader!../../../shared/variables.css';
+import bootstrapStyles from '!!raw-loader!../../../shared/themed-bootstrap.css';
 
 const template = document.createElement('template');
