+++ src/stable/stories/infobutton.stories.js
@@ -1,11 +1,11 @@
 import { html } from 'lit-html';
-import '../components/atoms/Icon/cod-icon';
-import '../components/atoms/InfoButton/cod-info-button';
+import '../../experimental/components/atoms/Icon/cod-icon';
+import '../components/InfoButton/cod-info-button';
 
 export default {
-  tags: ['experimental'],
+  tags: ['stable'],
   component: 'cod-info-button',
-  title: 'Experimental/Info Button',
+  title: 'Components/Info Button',
   // 👇 Creates specific argTypes
   argTypes: {
     titlePrimary: {
